Projects/04-Networking/2-Queues/api/requirements.txt
fastapi
uvicorn
pika
psycopg2-binary
python-multipart
